feat: adds library_name to vivado_library
By default, a vivado_library's name is simply the name of the target.

However, since a typical design will use multiple tools, defining the
tools' rules in the same BUILD file means we must adjust target names,
to allow them to coexist. However, we can not change the library names
in source code, so we are painted into a corner.

I added `vivado_library.library_name` parameter to allow this to be
modified.